hotshop/requirements.txt
asn1crypto==0.23.0
attrs==17.3.0
Automat==0.6.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.3
cssselect==1.0.1
Django==1.10.8
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-haystack==2.6.1
django-redis==4.8.0
django-rest-swagger==2.1.2
django-reversion==2.0.12
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
exrex==0.10.5
future==0.16.0
geoip2==2.6.0
gunicorn==19.7.1
httplib2==0.10.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itypes==1.1.0
jieba==0.39
Jinja2==2.10
lxml==4.1.1
Markdown==2.6.9
MarkupSafe==1.0
maxminddb==1.3.0
mysqlclient==1.3.12
olefile==0.44
openapi-codec==1.3.2
parsel==1.2.0
Pillow==4.3.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
pycryptodome==3.4.7
Pygments==2.2.0
PyJWT==1.5.3
PyMySQL==0.8.0
pyOpenSSL==17.3.0
pytz==2017.3
qiniu==7.2.0
queuelib==1.4.2
raven==6.4.0
redis==2.10.6
requests==2.18.4
service-identity==17.0.0
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
w3lib==1.18.0
Whoosh==2.7.4
XlsxWriter==1.0.2
xlwt==1.3.0
zope.interface==4.4.3
